% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LigandReceptorPairs.R
\name{LigandReceptorPairsTable}
\alias{LigandReceptorPairsTable}
\title{Create table with ligand-receptor pairs between cells.
This function combines gene expression output data from SEURAT with
a database of Ligand-Receptor pairs to identify potential
cell-cell interactions in a given dataset.}
\usage{
LigandReceptorPairsTable(
  ncells,
  celltypelabels,
  seuratDEGS,
  LRdatabase,
  subsetgenes = seuratDEGS$gene
)
}
\arguments{
\item{ncells}{Number of cell types in annotated seurat object}

\item{celltypelabels}{Vector containing labels for ncells}

\item{seuratDEGS}{Exact output from Seurat::FindAllMarkers function. Must contain a 'gene' column}

\item{LRdatabase}{Table of ligand-receptor pairs}

\item{subsetgenes}{Vector containing a subset of genes}
}
\description{
Create table with ligand-receptor pairs between cells.
This function combines gene expression output data from SEURAT with
a database of Ligand-Receptor pairs to identify potential
cell-cell interactions in a given dataset.
}
