% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRdatabase.R
\docType{data}
\name{LRdatabase}
\alias{LRdatabase}
\title{LRdatabase.}
\format{
A data frame with 2557 rows and 5 variables:
\describe{
  \item{Pair}{Gene symbols of ligand-receptor pair}
  \item{Ligand}{Gene symbol of Ligand}
  \item{Ligand.name}{Gene name of ligand}
  \item{Receptor}{Gene symbol of receptor}
  \item{Receptor.name}{Gene name of receptor}
}
}
\source{

}
\usage{
LRdatabase
}
\description{
A dataset containing 2557 ligand-receptor pairs
}
\keyword{datasets}
