% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_dataset.R
\docType{data}
\name{test_dataset}
\alias{test_dataset}
\title{test_dataset.}
\format{
A data frame with 12814 rows and 7 variables:
\describe{
  \item{p_val}{p value}
  \item{avg_logFC}{log Fold Change}
  \item{pct.1}{}
  \item{pct.2}{}
  \item{p_val_adj}{Adjusted p value}
  \item{cluster}{Cluster ID}
  \item{gene}{Gene name}
}
}
\source{
\url{}
}
\usage{
test_dataset
}
\description{
Example dataset containing gene expression values across 23 cell types
obtained using FindAllMarkers function from SEURAT v3.
}
\keyword{datasets}
