% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LigandReceptorPairs.R
\name{PairsPlot}
\alias{PairsPlot}
\title{Chord Plot of ligand-receptor interactions
This function combines gene expression output data from SEURAT with
a database of Ligand-Receptor pairs to identify potential
cell-cell interactions in a given dataset and uses the circlize package
to produce a chord plot representative of those interactions}
\usage{
PairsPlot(
  seuratDEGS,
  LRdatabase,
  cellcolors,
  subsetgenes = seuratDEGS$gene,
  from = celltypelabels,
  to = celltypelabels
)
}
\arguments{
\item{seuratDEGS}{Exact output from Seurat::FindAllMarkers function. Must contain a 'gene' column}

\item{LRdatabase}{Table of ligand-receptor pairs}

\item{cellcolors}{vector of color names or codes for each cell type}

\item{subsetgenes}{Vector containing a subset of genes}

\item{from}{vector of cell type names to subset outgoing interactions. Default is all cells.}

\item{to}{vector of cell type names to subset incoming interactions. Default is all cells.}
}
\description{
Chord Plot of ligand-receptor interactions
This function combines gene expression output data from SEURAT with
a database of Ligand-Receptor pairs to identify potential
cell-cell interactions in a given dataset and uses the circlize package
to produce a chord plot representative of those interactions
}
